+++ Changelog.md
@@ -8,10 +8,19 @@ CHANGELOG
 
 ### ??/??/2021 2.10
 
-- Add ICountable-like interfaces with the '(u)long' underlying type.
-- Add WinCopies.Collections.Generic.IReadOnlyList2\<out T> interface.
-- Add WinCopies.Util.ThrowHelper.GetOrThrowIfDisposed static method.
-- Add WinCopies.Util.DotNetFix.IDisposable.GetOrThrowIfDisposed\<T> default implementation.
+- Additions:
+	- Interfaces:
+		- ICountable-like interfaces with the '(u)long' underlying type.
+		- WinCopies.Collections.Generic.IReadOnlyList2\<out T> interface.
+	- Methods:
+		- WinCopies.Util.ThrowHelper.GetOrThrowIfDisposed static method.
+		- WinCopies.Util.DotNetFix.IDisposable.GetOrThrowIfDisposed\<T> default implementation.
+	- Classes:
+		- CustomEnumeratorEnumerable/Provider\<TItems, TEnumerator>
+		- ToStringConverter
+		- SelectorDictionary
+		- WinCopies.Delegates and WinCopies.Bool that are classes with static methods that can be considered as default implementation for some usual delegates.
+	- ApplicationCommands.Reset command.
 
 #### WinCopies.Collections